// apps/web/src/lib/api/zod/models/ruleCondition.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Sure API
 * OpenAPI documentation generated from executable request specs.
 * OpenAPI spec version: v1
 */
import * as zod from 'zod';

export type RuleCondition = {
  id: string;
  condition_type: string;
  operator: string;
  /** @nullable */
  value?: string | null | undefined;
  sub_conditions: RuleCondition[];
  created_at: string;
  updated_at: string;
};

export const RuleCondition: zod.ZodType<RuleCondition> = zod.object({
  "id": zod.uuid(),
  "condition_type": zod.string(),
  "operator": zod.string(),
  "value": zod.string().nullish(),
  "sub_conditions": zod.array(zod.lazy(() => RuleCondition)),
  "created_at": zod.iso.datetime({"offset":true}),
  "updated_at": zod.iso.datetime({"offset":true})
});

export type RuleConditionOutput = zod.output<typeof RuleCondition>;